Apex status tool
Neptune Fusion's cloud authentication creates complex cookie management
requirements. The POST login returns 3 cookies (2x connect.sid with
different paths, 1x remember_me) that are all required for subsequent
API requests.
The authenticator returns an array of HTTP::Cookie objects instead of a
hash because it's the easiest way for it to be reused by other services
that will make calls to fusion. Also we will almost certainly learn more
when we make other services that interact with Apex Fusion so I felt it
was best to focus on features for now and we'll probably make this nicer
down the road